+++ apps/sim/app/api/table/[tableId]/import/route.test.ts
@@ -45,13 +45,13 @@ vi.mock('@/app/api/table/utils', async () => {
 })
 
 /**
- * The route imports `importAppendRows` / `importReplaceRows` from the barrel,
- * which forwards them from `./service`. These functions own the import
- * transaction (column adds + row writes); mocking the service module replaces
- * them without touching the other real helpers (`coerceRowsForTable`,
- * `createCsvParser`, etc.) exported through the barrel.
+ * The route imports `importAppendRows` / `importReplaceRows` from
+ * `@/lib/table/import-data`. These functions own the import transaction (column
+ * adds + row writes); mocking that module replaces them without touching the
+ * other real helpers (`coerceRowsForTable`, `createCsvParser`, etc.) exported
+ * through the barrel.
  */
-vi.mock('@/lib/table/service', () => ({
+vi.mock('@/lib/table/import-data', () => ({
   importAppendRows: mockImportAppendRows,
   importReplaceRows: mockImportReplaceRows,
 }))
