+++ apps/sim/lib/table/index.ts
@@ -10,6 +10,7 @@ export * from './column-keys'
 export * from './columns/service'
 export * from './constants'
 export * from './import'
+export * from './import-data'
 export * from './jobs/service'
 export * from './llm'
 export * from './query-builder'
